requirements.txt
aiostream~=0.4.4
autobahn~=22.2.2
asyncio-mqtt~=0.11.0
beanie~=1.8.12
python-decouple~=3.5
simplejson~=3.17.6
-e git+https://github.com/PatrickBaus/pyAsyncPrologixGpib#egg=prologix_gpib_async
-e git+https://github.com/PatrickBaus/pyAsyncHP3478A#egg=hp3478a_async
-e git+https://github.com/PatrickBaus/TinkerforgeAsync@generators#egg=tinkerforge_async
-e git+https://github.com/PatrickBaus/LabNode#egg=labnode_async
